// packages/ui/src/types/sanity.ts
export interface SanityDocument {
  _id: string;
  _type: string;
  _rev?: string;
  _createdAt?: string;
  _updatedAt?: string;
}

export interface SanitySlug {
  _type: 'slug';
  current: string;
}

export interface SanityImage {
  _type: 'image';
  asset: {
    _id?: string;
    _ref: string;
    _type: 'reference';
    url?: string;
    metadata?: {
      dimensions?: {
        aspectRatio?: number;
        height?: number;
        width?: number;
      };
      lqip?: string;
      blurHash?: string;
    }
  };
  crop?: {
    _type: 'sanity.imageCrop';
    bottom: number;
    left: number;
    right: number;
    top: number;
  };
  hotspot?: {
    _type: 'sanity.imageHotspot';
    height: number;
    width: number;
    x: number;
    y: number;
  };
  alt?: string;
}
// ...
export interface SanityPageContent {
  _key?: string;
  _type: string;
  [key: string]: any; // Additional properties based on block type
  theme?: 'light' | 'dark';
  responsivePadding?: any; // Responsive padding configuration
  htmlId?: string;
  backgroundImage?: SanityImage;
  minHeight?: 'none' | 'sm' | 'md' | 'lg';
}
// ...
export interface SanityCompany extends SanityDocument {
  _type: 'company';
  name: string;
  logo?: SanityImage;
  logoOnLight?: SanityImage;
  logoOnDark?: SanityImage;
  logomarkOnLight?: SanityImage;
  logomarkOnDark?: SanityImage;
  website?: string;
};